refactor: EmbeddedRedisConfig 내용 RedissonConfig으로 합치기
 - profile local or test일 때 redis embedded 환경으로 사용하는 내용 RedissonConfig 클래스 내로 옮김
 - test 환경 설정 시 둘 다 classes에 포함해야 되는 번거로움 줄이기 위함
