+++ src/test/kotlin/com/mojh/cms/infra/RedissonTest.kt
@@ -1,7 +1,6 @@
 package com.mojh.cms.infra
 
 import com.mojh.cms.common.BaseTest
-import com.mojh.cms.common.config.EmbeddedRedisConfig
 import com.mojh.cms.common.config.RedissonConfig
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -10,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest
 import java.util.concurrent.TimeUnit
 
 @BaseTest
-@SpringBootTest(classes = [RedissonConfig::class, EmbeddedRedisConfig::class])
+@SpringBootTest(classes = [RedissonConfig::class])
 class RedissonTest(private val redisson: RedissonClient) {
 
     @Test
